+++ docs/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Changelog
 
+### Release v1.1.0 `10/11/2021`
+
+- Added [PaginatedResults](../src/Entity/PaginatedResults.php) to better structure the repository results for `fetchAll()` requests.
+- Updated unit tests to cover this new functionality.
+
 ### Release v1.0.0 `09/11/2021`
 
 Initial application layout.